Added Docker support
You need Docker version 17 or later to execute that.

This is part of an attempt to automate the creation of replicas. Now with just a simple command is possible to start 3 groups of replicas (12 replicas in total) running BFT-SMART with different broadcast domain. You can choose between one of the two commands:

	- docker-compose up -d --build --no-deps
	- make up