Added PoC
In this commit we added the required files of the Proof of Concept of
our non-genuine atomic multicast system.

Based on official documentation of BFT-SMART, this application (trought
the Proxy.java) performs the message forwarding according of the target
group.

Now that we know it works, more work ahead awaits for us.