Added Client/Proxy Application
Now is possible to deploy multiple client and proxy application, so if you
want deploy a proxy/client that connects in replicas of group 0:

	$ docker-compose -f docker-compose.proxy.yml up --build --no-deps -d group_0_proxy_0

And to perform a simple benchmark in the cluster:
	$ java -cp '../lib/*:bin/*' ch.usi.dslab.ceolin.bftswarm.Client <cliend id> <num groups> <num ops> <local config> <global config>

Remember to use a unique <client id> to connect into the
proxy/replicas. If a number greater than 1 is set in <num groups>, the
messages will be received by all replicas of the system, so expect a
poor performance here.