Quick test to track and compare consensus id
This is a initial version and not really well tested commit that changes the behavior of replicas, both local and global replicas. Now we try to use the information provided by the MessageContext class to check if the order of messages match.

To do that we save in the global replicas the last consensus id sent to each group. When a new message is sent, in addition to the current consensus id, the last consensus id sent to this group is also encapsulated. With this information, the local replica is able to verify the order or if any message has been lost.